+++ app.js
@@ -2,7 +2,7 @@ const express = require("express");
 const cors = require("cors");
 const logger = require("morgan");
 
-const contactsRouter = require("./routes/api/contacts");
+const contactsRouter = require("./routes/contacts");
 
 const app = express();
