+++ controllers/contacts/index.js
@@ -1,13 +1,3 @@
-const getAll = require("./getAllContacts");
-const getContactById = require("./getContactById");
-// const postContact = require("./postContact");
-const deleteContact = require("./deleteContact");
-const changeContact = require("./changeContact");
+const getAll = require("./getAll");
 
-module.exports = {
-  getAll,
-  getContactById,
-  // postContact,
-  deleteContact,
-  changeContact,
-};
+module.exports = { getAll };
