+++ middlewares/index.js
@@ -1,0 +1,3 @@
+const ctrlWrapper = require("./ctrlWrapper");
+const validation = require("./validation");
+module.exports = { ctrlWrapper, validation };
