+++ models/contact.js
@@ -25,4 +25,5 @@ const contactSchema = Schema({
 });
 
 const Contact = model("contact", contactSchema);
+
 module.exports = Contact;
