--- routes/api/contacts.js
@@ -1,16 +1,0 @@
-const express = require("express");
-const router = express.Router();
-
-const ctrl = require("../../controllers");
-
-router.get("/", ctrl.getAll);
-
-router.get("/:contactId", ctrl.getContactById);
-
-// router.post("/", ctrl.postContact);
-
-router.delete("/:contactId", ctrl.deleteContact);
-
-router.put("/:contactId", ctrl.changeContact);
-
-module.exports = router;
