--- schemas/index.js
@@ -1,3 +1,0 @@
-const Product = require("./contacts");
-
-module.exports = Product;
